# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Monitoring and operations API for Redis Sentinel clusters on Kubernetes.",
        "title": "Redis Sentinel Admin API",
        "contact": {
            "name": "Internal Infra Team"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/replication/lag": {
            "get": {
                "description": "Returns the replication offset lag in bytes for every replica, plus a ring-buffer trend of recent samples and a promotion score (0–100, higher = better failover candidate).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replication"
                ],
                "summary": "Replication lag per replica",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_replication.ReplicaLag"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "502": {
                        "description": "Node unreachable",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "No master",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/replication/resync-stats": {
            "get": {
                "description": "Returns full-resync counters and backlog utilisation per node. Raises an alert when the backlog exceeds 90% capacity and includes a sizing advisory.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replication"
                ],
                "summary": "Replication backlog and resync stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_replication.ResyncReport"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "502": {
                        "description": "Node unreachable",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "No master",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_duydinhle_redis-sentinel-admin_internal_api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIError"
                }
            }
        },
...
        "github_com_duydinhle_redis-sentinel-admin_internal_replication.LagSample": {
            "type": "object",
            "properties": {
                "at": {
                    "type": "string"
                },
                "lag_bytes": {
                    "type": "integer"
                }
            }
        },
        "github_com_duydinhle_redis-sentinel-admin_internal_replication.ReplicaLag": {
            "type": "object",
            "properties": {
                "is_caught_up": {
                    "description": "lag \u003c 1 MiB",
                    "type": "boolean"
                },
                "lag_bytes": {
                    "type": "integer"
                },
                "lag_trend": {
                    "description": "up to 10 most recent samples",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_replication.LagSample"
                    }
                },
                "master_offset": {
                    "type": "integer"
                },
                "node_addr": {
                    "type": "string"
                },
                "promotion_score": {
                    "description": "0–100, higher = better failover candidate",
                    "type": "number"
                },
                "replica_offset": {
                    "type": "integer"
                }
            }
        },
        "github_com_duydinhle_redis-sentinel-admin_internal_replication.ResyncReport": {
            "type": "object",
            "properties": {
                "advisory": {
                    "type": "string"
                },
                "backlog_active_size_bytes": {
                    "type": "integer"
                },
                "backlog_alert": {
                    "description": "active \u003e 90% of backlog size",
                    "type": "boolean"
                },
                "backlog_size_bytes": {
                    "type": "integer"
                },
                "node_addr": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "total_resyncs": {
                    "type": "integer"
                }
            }
        },
...
    }
}